Proj1. Q5. Good state evaluation function. Fixing issues:
1. Power-food positions are considered - will head for these first-hand,
as that turns ghosts fleeing, and eating a fleeing ghost is +200 to state.

2. Evaluates position to closest food using proper A* tree-searching,
instead of a Manhattan-distance estimate (which doesn’t take walls into account).
This solves the issue of being stuck behind the walls: the move along the wall
takes you closer to the closest food, and not further.